tried to add a few optimizations and did first speed tests against libc (by porting my tests to malloc/free/realloc)

+++ internal/virtalloc/checksum.h
@@ -3,6 +3,6 @@
 
 #include <stddef.h>
 
-size_t get_checksum(size_t length, const char data[static length]);
+int get_checksum(size_t length, const char data[static length]);
 
 #endif
